refactor: externalize StandaloneTopBar component
